First pass at drawRect.  No love with YValueForX:.

+++ GraphCalc/GraphView.h
@@ -12,7 +12,7 @@
 
 @protocol GraphViewDelegate
 @property (nonatomic) int scale;
-- (double)YValueForX:(int)x;
+- (double)YValueForX:(double)x;
 @end
